FIXING
I did a lot of bug fixes and tried and tried and I've fixed most of the problems that had to do with builds and why it was hard to move and look around and pick stuff up. Now I just need to make some minore adjustments tomorrow and I'll be set to film what little we have to show.