FInal push from me
I made the adjustments that I could do and have now filmed footage from our little start menu. The bottle didn't ork in the end and I have no idea why but at least I've tried. I guess my lingering regret is not being able to put together all the scenes so we had a cohecent game but it is what it is. This project has been fun :)